[breadboard -> breadboard-web] move UI
