+++ seeds/breadboard-web/src/ui/output.ts
@@ -4,8 +4,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { OutputValues, type Schema } from "../types.js";
-import { StreamCapabilityType } from "../stream.js";
+import {
+  OutputValues,
+  StreamCapabilityType,
+  type Schema,
+} from "@google-labs/breadboard";
 
 export type OutputArgs = Record<string, unknown> & {
   schema: Schema;
